added logic to append to existing new_name_fasta if more than one record, otherwise write new fasta.

+++ workflow/scripts/rename_fasta.py
@@ -12,7 +12,13 @@ with open(snakemake.input.samp_fasta) as original_file:
         name_change_df.loc[i] = [ snakemake.input.samp_fasta , seq.id, snakemake.output.new_fa ,f"{snakemake.wildcards.sm}__{snakemake.wildcards.h}__{i}"]
         seq.id = name_change_df.loc[i, "new_name"]
         seq.description = "" #description added to end of header. was old name previously.
-    with open(snakemake.output.new_fa, "w") as output_handle:
-        SeqIO.write(records, output_handle, "fasta")
+
+    if(i > 0): # if more than one record then append to existing fasta. Otherwise if just one record write new file.
+        with open(snakemake.output.new_fa, "a") as output_handle:
+            SeqIO.write(records, output_handle, "fasta")
+    else:
+        with open(snakemake.output.new_fa, "w") as output_handle:
+            SeqIO.write(records, output_handle, "fasta")
+
 os.system(f"samtools faidx {snakemake.output.new_fa}")
 name_change_df.to_csv( snakemake.output.old_new_name_map , sep = "\t", header = True, index = False)
